feat: move meetup event feature into its own schematic
This change pulls the meetup event feature into its own schematic so that the ng-add schematic will
be smaller and easier to maintain in the long run. It also adds some script updates to the user's
`package.json` file, adds a new sample site background image, and updates the readme.