refactor(dependencies): refactctor project setup schematic
This change updates the way the schematics handle dependencies. Previously, external schematics that
Community Schematics depended on were listed as required dependencies for the project. For example,
this meant that if a user added Community Schematics but didn't intend to deploy to Azure, the Azure
deploy schematics would still be downloaded and installed as part of the `ng add` process. This
change now installs only the dependencies that are required as part of the `ng add` process at the
time of execution instead of keeping them as part of the package overall.